update: デザイン変更

+++ vihicle/model.go
@@ -4,7 +4,7 @@ import "github.com/mistralll/nenpi/refueling"
 
 type Vihicle struct {
 	Title      string
-	AvgMileage int64
+	AvgMileage float64
 	Refuelings []refueling.Refueling
 }
